+++ changelog.md
@@ -5,6 +5,7 @@
  - Handle the permissions request
  - Report issue rendering ListItem with secondaryText to Material-UI
 ### v0.3.0
+* 20170613 | F | Fixed reference on Wallpaper component
 * 20170612 | P | Minor improvements to Tiles
 * 20170609 | P | Added tiles
 * 20170608 | P | Added ThumbnailsProvider